requirements.txt
rdkit~=2023.9.1
numpy~=1.26.1
pandas~=2.1.2
matplotlib~=3.8.1
scikit-learn~=1.3.2
mordred~=1.2.0
molml~=0.9.0
tensorflow~=2.14.0
keras~=2.14.0
seaborn~=0.13.0
streamlit~=1.30.0
Pillow~=10.1.0
requests~=2.31.0
streamlit-option-menu
streamlit-antd-components
docxtpl~=0.16.7
lightgbm~=4.1.0
joblib~=1.3.2
chembl-webresource-client==0.10.8
streamlit-ketcher
py2opsin~=1.0.6